% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv-data.R, R/adv2-data.R, R/advpres-data.R
\docType{data}
\name{adv}
\alias{adv}
\alias{adv2}
\alias{advpres}
\title{Dominance sequence from Albers and de Vries (2001)}
\format{Fictional example of an interaction sequence, with 33 interactions between 7 individuals.}
\usage{
data(adv)

data(adv2)

data(advpres)
}
\description{
Dominance sequence from Albers and de Vries (2001)

Dominance sequence from Albers and de Vries (2001) with added information about interaction type and whether interaction ended in a draw

Fictional presence data Albers and de Vries (2001)
}
\examples{
data(adv)
data(adv2)
data(advpres)
}
\references{
Albers, P. C. H. & de Vries, H. 2001. Elo-rating as a tool in the sequential estimation of dominance strengths. Animal Behaviour, 61, 489-495. (\href{https://dx.doi.org/10.1006/anbe.2000.1571}{DOI: 10.1006/anbe.2000.1571})

Albers, P. C. H. & de Vries, H. 2001. Elo-rating as a tool in the sequential estimation of dominance strengths. Animal Behaviour, 61, 489-495. (\href{https://dx.doi.org/10.1006/anbe.2000.1571}{DOI: 10.1006/anbe.2000.1571})

Albers, P. C. H. & de Vries, H. 2001. Elo-rating as a tool in the sequential estimation of dominance strengths. Animal Behaviour, 61, 489-495. (\href{https://dx.doi.org/10.1006/anbe.2000.1571}{DOI: 10.1006/anbe.2000.1571})
}
\keyword{datasets}

