% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advpres-data.R
\docType{data}
\name{advpres}
\alias{advpres}
\title{Fictional presence data for Albers and de Vries (2001)}
\format{Fictional example of an interaction sequence, with 33 interactions between 7 individuals.}
\usage{
data(advpres)
}
\description{
Fictional presence data Albers and de Vries (2001)
}
\examples{
data(advpres)
}
\references{
Albers, P. C. H. & de Vries, H. 2001. Elo-rating as a tool in the sequential estimation of dominance strengths. Animal Behaviour, 61, 489-495. (\href{https://dx.doi.org/10.1006/anbe.2000.1571}{DOI: 10.1006/anbe.2000.1571})
}
\keyword{datasets}
