% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baboons-data.R
\docType{data}
\name{baboons1}
\alias{baboons1}
\alias{baboons2}
\alias{baboons3}
\alias{baboons4}
\alias{baboons5}
\alias{baboons}
\title{Baboon dominance sequences}
\format{
Data sets of 5 groups of baboons, with date, winner and loser columns
}
\usage{
baboons1
}
\description{
Baboon dominance sequences
}
\details{
The exact dates of the interactions were not given in the actual online data sets, so I set them to roughly match the data collection period presented in the actual paper (1996 - 2011)
}
\examples{
data(baboons1)
}
\references{
\insertRef{franz2015a}{EloRating}

\insertRef{franz2015c}{EloRating}
}
\keyword{datasets}
