% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonobo-data.R
\docType{data}
\name{bonobos}
\alias{bonobos}
\title{Dominance matrix from de Vries et al. 2006}
\format{Integer matrix, with column and row names. Winners in rows and losers in columns.}
\usage{
data(bonobos)
}
\description{
Dominance matrix of seven bonobos
}
\examples{
data(bonobos)
}
\references{
de Vries, H., Stevens, J. M. G. & Vervaecke, H. 2006. Measuring and testing the steepness of dominance hierarchies. Animal Behaviour, 71, 585-592. (\href{dx.doi.org/10.1016/j.anbehav.2005.05.015}{DOI: 10.1016/j.anbehav.2005.05.015})
}
\keyword{datasets}

