% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devries98-data.R
\docType{data}
\name{devries98}
\alias{devries98}
\title{Dominance matrix from de Vries (1998)}
\format{Named integer matrix.}
\usage{
data(devries98)
}
\description{
Fictional dominance matrix from de Vries (1998) from 10 individuals.
}
\examples{
data(devries98)
}
\references{
de Vries, H. 1998. Finding a dominance order most consistent with a linear hierarchy: a new procedure and review. Animal Behaviour, 55, 827-843. (\href{https://dx.doi.org/10.1006/anbe.1997.0708}{DOI: 10.1006/anbe.1997.0708})
}
\keyword{datasets}

