% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dommats-data.R
\docType{data}
\name{dommats}
\alias{dommats}
\title{Example dominance matrices}
\format{
A named list with dominance matrices:
\itemize{
\item \code{badgers}: 7 badgers (Hewitt et al 2009, Fig. A1 PO2004)
\item \code{squirrels}: 8 squirrels (Farentinos 1972, Table 1C)
\item \code{elephants}: 7 elephants (Archie et al 2006, Fig. 2, JA)
}
}
\usage{
dommats
}
\description{
Example dominance matrices
}
\examples{
data(dommats)
}
\references{
\insertRef{farentinos1972}{EloRating}

\insertRef{archie2006}{EloRating}

\insertRef{hewitt2009}{EloRating}
}
\keyword{datasets}
