% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISI-diffmat.R
\name{.diffmat}
\alias{.diffmat}
\title{difference matrix}
\usage{
.diffmat(mat)
}
\arguments{
\item{mat}{square interaction matrix with winner in rows and losers in columns, for example the output from \code{\link{creatematrix}}}
}
\value{
a matrix with ranking differences assuming that the matrix reflects the order. This information is contained in the upper triangle of the returned matrix.
}
\description{
difference matrix
}
\details{
helper function for \code{\link{ISI}}
}
\examples{
data(bonobos)
EloRating:::.diffmat(bonobos)

}
\author{
Christof Neumann
}
