% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISI-incon.R
\name{.incon}
\alias{.incon}
\title{number of inconsistencies}
\usage{
.incon(mat)
}
\arguments{
\item{mat}{square interaction matrix with winner in rows and losers in columns, for example the output from \code{\link{creatematrix}}}
}
\value{
integer, the number of inconsistencies in the matrix
}
\description{
calculate number of inconsistencies
}
\examples{
data(bonobos)
EloRating:::.incon(bonobos)

}
\references{
\insertRef{devries1998}{EloRating}
}
\author{
Christof Neumann
}
