% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS.R
\name{DS}
\alias{DS}
\title{David's score}
\usage{
DS(interactionmatrix, prop = c("Dij", "Pij"))
}
\arguments{
\item{interactionmatrix}{square interaction matrix with winner in rows and losers in columns, for example the output from \code{\link{creatematrix}}}

\item{prop}{the type of dyadic win proportion to be use. By default corrected for number of interactions in a dyad (\code{prop="Dij"}), otherwise the raw proportion (\code{prop="Pij"})}
}
\value{
a data.frame with columns ID, DS (David's scores) and normDS (normalized David's scores)
}
\description{
calculate David's scores from an interaction matrix
}
\examples{
data(bonobos)
DS(bonobos)

}
\references{
David, H. A. 1987. Ranking from unbalanced paired-comparison data. Biometrika, 74, 432-436. (\href{https://dx.doi.org/10.1093/biomet/74.2.432}{DOI: 10.1093/biomet/74.2.432})

de Vries, H., Stevens, J. M. G. & Vervaecke, H. 2006. Measuring and testing the steepness of dominance hierarchies. Animal Behaviour, 71, 585-592. (\href{https://dx.doi.org/10.1016/j.anbehav.2005.05.015}{DOI: 10.1016/j.anbehav.2005.05.015})

Gammell, M. P., de Vries, H., Jennings, D. J., Carlin, C. M. and Hayden, T. J. 2003. David's score: a more appropriate dominance ranking method than Clutton-Brock et al.'s index. Animal Behaviour, 66, 601-605. (\href{https://dx.doi.org/10.1006/anbe.2003.2226}{DOI: 10.1006/anbe.2003.2226})
}
\author{
Christof Neumann
}
