% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_dom.R
\name{dyadic_dom}
\alias{dyadic_dom}
\title{dyadic dominance relations}
\usage{
dyadic_dom(winner, loser, Date = NULL, daterange = NULL)
}
\arguments{
\item{winner}{character or factor with winner}

\item{loser}{character or factor with winner}

\item{Date}{not yet implemented}

\item{daterange}{not yet implemented}
}
\value{
a data.frame with one row per dyad
}
\description{
dyadic dominance relations
}
\examples{
xdata <- randomsequence(nID = 5, avgIA = 10, reversals = 0.1)$seqdat
dyadic_dom(xdata$winner, xdata$loser)
}
