\name{Elo2pack-package}
\alias{Elo2pack-package}
\alias{Elo2pack}
\docType{package}
\title{
\packageTitle{Elo2pack}
}
\description{
\packageDescription{Elo2pack}
}
\details{
An overview of how to use the package, including the most important functions.
}
\author{
\packageAuthor{Elo2pack}

Maintainer: \packageMaintainer{Elo2pack}
}
\references{
Elo 1978
Albers and de Vries 2001
}
\keyword{ package }

\examples{
rnorm(10)
}
