% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISI-incontable.R
\name{incontable}
\alias{incontable}
\title{number and strength of inconsistencies}
\usage{
incontable(mat)
}
\arguments{
\item{mat}{square interaction matrix with winner in rows and losers in columns, for example the output from \code{\link{creatematrix}}}
}
\value{
data frame with inconsistencies and their strength
}
\description{
calculate number and strength of inconsistencies
}
\examples{
data(bonobos)
incontable(bonobos)

}
\references{
de Vries, H. 1998. Finding a dominance order most consistent with a linear hierarchy: a new procedure and review. Animal Behaviour, 55, 827-843. (\href{https://dx.doi.org/10.1006/anbe.1997.0708}{DOI: 10.1006/anbe.1997.0708})
}
\author{
Christof Neumann
}
