% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISI.R
\name{ISI}
\alias{ISI}
\title{strength of inconsistencies}
\usage{
ISI(mat, runs = 5000)
}
\arguments{
\item{mat}{square interaction matrix with winner in rows and losers in columns, for example the output from}

\item{runs}{numeric, number of iterations}
}
\value{
integer, the summed strength of inconsistencies in the matrix
}
\description{
calculate strength of inconsistencies
}
\details{
The number of interations is set substantially higher than what was suggested in the de Vries' 1998 paper, because the algorithm here is less efficient.

The I&SI algorithm (c.f. de Vries 1998) does not necessarily result in a unique order (see example below). If such a case occurs, all (equally good) solutions are returned as a list.
}
\examples{
\dontrun{
 data(devries98)
 h.index(devries98)
 ISI(devries98)

 ##
 data(adv)
 # SEQ <- elo.seq(winner=adv$winner, loser=adv$loser, Date=adv$Date)
 # mat <- creatematrix(SEQ)
 # res <- ISI(mat)
 # note that this matrix is not sufficiently linear to justify such ordering
 # h.index(mat)}


}
\author{
Christof Neumann
}
\references{
de Vries, H. 1998. Finding a dominance order most consistent with a linear hierarchy: a new procedure and review. Animal Behaviour, 55, 827-843. (\href{https://dx.doi.org/10.1006/anbe.1997.0708}{DOI: 10.1006/anbe.1997.0708})
}

