% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.seqchecknopres.r
\name{print.seqchecknopres}
\alias{print.seqchecknopres}
\title{prints its argument}
\usage{
\method{print}{seqchecknopres}(x, ...)
}
\arguments{
\item{x}{result from \code{\link{seqcheck}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
prints its argument
}
\examples{
data(adv)
print(seqcheck(winner = adv$winner, loser = adv$loser, Date = adv$Date))
}
\author{
Christof Neumann
}
