% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv-data.R
\docType{data}
\name{adv}
\alias{adv}
\title{Dominance sequence from Albers and de Vries (2001)}
\format{Fictional example of an interaction sequence, with 33 interactions between 7 individuals.}
\usage{
data(adv)
}
\description{
Dominance sequence from Albers and de Vries (2001)
}
\examples{
data(adv)
}
\references{
Albers, P. C. H. & de Vries, H. 2001. Elo-rating as a tool in the sequential estimation of dominance strengths. Animal Behaviour, 61, 489-495. (\href{https://dx.doi.org/10.1006/anbe.2000.1571}{DOI: 10.1006/anbe.2000.1571})
}
\keyword{datasets}
