% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv-data.R
\docType{data}
\name{adv}
\alias{adv}
\title{Dominance sequence from Albers and de Vries (2001)}
\format{
Fictional example of an interaction sequence, with 33 interactions between 7 individuals.
}
\usage{
data(adv)
}
\description{
Dominance sequence from Albers and de Vries (2001)
}
\examples{
data(adv)
}
\references{
\insertRef{albers2001}{EloRating}
}
\keyword{datasets}
