% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv2-data.R
\docType{data}
\name{adv2}
\alias{adv2}
\title{Dominance sequence from Albers and de Vries (2001)}
\format{
Fictional example of an interaction sequence, with 33 interactions between 7 individuals.
}
\usage{
data(adv2)
}
\description{
Dominance sequence from Albers and de Vries (2001) with added information about interaction type and whether interaction ended in a draw
}
\examples{
data(adv2)
}
\references{
\insertRef{albers2001}{EloRating}
}
\keyword{datasets}
