% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advpres-data.R
\docType{data}
\name{advpres}
\alias{advpres}
\title{Fictional presence data for Albers and de Vries (2001)}
\format{
Fictional example of an interaction sequence, with 33 interactions between 7 individuals.
}
\usage{
data(advpres)
}
\description{
Fictional presence data for Albers and de Vries (2001)
}
\examples{
data(advpres)
}
\references{
\insertRef{albers2001}{EloRating}
}
\keyword{datasets}
