\name{bonobos}
\alias{bonobos}
\docType{data}
\title{
Dominance matrix from de Vries et al. 2006
}
\description{
Dominance matrix of seven bonobos
}
\usage{data(bonobos)}
\format{
  Dominance matrix of seven bonobos
}
\source{
de Vries, H., Stevens, J. M. G. & Vervaecke, H. 2006. Measuring and testing the steepness of dominance hierarchies. Animal Behaviour, 71, 585-592.
}
\references{
de Vries, H., Stevens, J. M. G. & Vervaecke, H. 2006. Measuring and testing the steepness of dominance hierarchies. Animal Behaviour, 71, 585-592.
}
\examples{
data(bonobos)
}
\keyword{datasets}
