% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonobo-data.R
\docType{data}
\name{bonobos}
\alias{bonobos}
\title{Dominance matrix from de Vries et al. 2006}
\format{
Integer matrix, with column and row names. Winners in rows and losers in columns.
}
\usage{
data(bonobos)
}
\description{
Dominance matrix of seven bonobos
}
\examples{
data(bonobos)
}
\references{
\insertRef{devries2006}{EloRating}
}
\keyword{datasets}
