% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devries98-data.R
\docType{data}
\name{devries98}
\alias{devries98}
\title{Dominance matrix from de Vries (1998)}
\format{
Named integer matrix.
}
\usage{
data(devries98)
}
\description{
Fictional dominance matrix from de Vries (1998) from 10 individuals.
}
\examples{
data(devries98)
}
\references{
\insertRef{devries1998}{EloRating}
}
\keyword{datasets}
