% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS.R
\name{DS}
\alias{DS}
\title{David's score}
\usage{
DS(interactionmatrix, prop = c("Dij", "Pij"))
}
\arguments{
\item{interactionmatrix}{square interaction matrix with winner in rows and losers in columns, for example the output from \code{\link{creatematrix}}}

\item{prop}{the type of dyadic win proportion to be use. By default corrected for number of interactions in a dyad (\code{prop="Dij"}), otherwise the raw proportion (\code{prop="Pij"})}
}
\value{
a data.frame with columns ID, DS (David's scores) and normDS (normalized David's scores)
}
\description{
calculate David's scores from an interaction matrix
}
\examples{
data(bonobos)
DS(bonobos)
DS(bonobos, prop = "Pij")

}
\references{
\insertRef{david1987}{EloRating}

\insertRef{gammell2003}{EloRating}

\insertRef{devries2006}{EloRating}
}
\author{
Christof Neumann
}
