% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.index.R
\name{h.index}
\alias{h.index}
\title{linearity indices}
\usage{
h.index(interactionmatrix, loops = 1000)
}
\arguments{
\item{interactionmatrix}{square interaction matrix with winner in rows and losers in columns, for example the output from \code{\link{creatematrix}}}

\item{loops}{numeric, the number of randomizations to perform (by default: 1000)}
}
\value{
a data.frame with with values for the number of individuals in the matrix (N), linearity indices (h, h' and expected h), p-value, number of randomizations, and number of unknown and tied relationships.
}
\description{
linearity indices
}
\details{
Note that the expected value of \emph{h} can also be calculated as 3/(N+1).
}
\examples{
data(bonobos)
h.index(bonobos)

}
\references{
\insertRef{appleby1983}{EloRating}

\insertRef{devries1995}{EloRating}
}
\author{
Christof Neumann
}
