% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.elo.R
\name{print.elo}
\alias{print.elo}
\title{prints its argument}
\usage{
\method{print}{elo}(x, ...)
}
\arguments{
\item{x}{result from \code{\link{elo.seq}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
prints its argument
}
\examples{
data(adv)
SEQ <- elo.seq(winner=adv$winner, loser=adv$loser, Date=adv$Date)
print(SEQ)
}
\author{
Christof Neumann
}
