% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sequencecheck.r
\name{print.sequencecheck}
\alias{print.sequencecheck}
\title{prints its argument}
\usage{
\method{print}{sequencecheck}(x, ...)
}
\arguments{
\item{x}{result from \code{\link{seqcheck}}}

\item{...}{further arguments passed to or from other methods (ignored here)}
}
\description{
prints its argument
}
\examples{
data(adv)
data(advpres)
print(seqcheck(winner = adv$winner, loser = adv$loser, Date = adv$Date,
                 presence = advpres))
}
\author{
Christof Neumann
}
