% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elo.R
\name{summary.elo}
\alias{summary.elo}
\title{summarize elo object}
\usage{
\method{summary}{elo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"elo"}, usually the result of a call to \code{\link{elo.seq}}}

\item{...}{further arguments passed to or from other methods (ignored)}
}
\description{
summarize elo object
}
\examples{
IA <- randomsequence()
SEQ <- elo.seq(winner=IA$seqdat$winner, loser=IA$seqdat$loser,
               Date=IA$seqdat$Date, draw=IA$seqdat$Draw,
               presence=IA$pres)
summary(SEQ)
}
\author{
Christof Neumann
}
