% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivity.R
\name{transitivity}
\alias{transitivity}
\title{triangle transitivity}
\usage{
transitivity(m, runs = 2000, returnfig = FALSE)
}
\arguments{
\item{m}{square dominance matrix}

\item{runs}{numeric, the number of runs for the randomization test}

\item{returnfig}{logical, should a figure be produced that shows the distribution of expectation}
}
\value{
a named vector of length four
}
\description{
triangle transitivity
}
\examples{
data(bonobos)
transitivity(bonobos)
}
\references{
\insertRef{shizuka2012}{EloRating}

\href{http://www.shizukalab.com/toolkits/sna/triangle-transitivity}{http://www.shizukalab.com/toolkits/sna/triangle-transitivity}
}
